Use BQM for Theta under in maxgap
